+++ game/scripts/run/coin.gd
@@ -1,6 +1,6 @@
 extends Area2D
 
-signal collected
+signal collected(at: Vector2)
 
 var _collected: bool = false
 
@@ -9,5 +9,5 @@ func collect() -> void:
 	if _collected:
 		return
 	_collected = true
-	collected.emit()
+	collected.emit(global_position)
 	queue_free()